dom/webgpu/mochitest/test_device_lost.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
  </head>
  <body>
    <script>
      ok(
        SpecialPowers.getBoolPref("dom.webgpu.enabled"),
        "Pref should be enabled."
      );

      const destroy_causes_lost = async function () {
        const adapter = await navigator.gpu.requestAdapter();
        ok(adapter !== undefined, "adapter !== undefined");
        const device = await adapter.requestDevice();
        ok(device !== undefined, "device !== undefined");

        const lostPromise = device.lost;
        device.destroy();
        const deviceLostReason = await lostPromise;

        is(
          deviceLostReason.reason,
          "destroyed",
          "Destroy reason should correspond to GPUDeviceLostReason.destroyed"
        );
        is(deviceLostReason.message, "", "Destroy message should be blank");
      };

      SimpleTest.waitForExplicitFinish();

      destroy_causes_lost()
        .catch(e => ok(false, `Unhandled exception ${e}`))
        .finally(() => SimpleTest.finish());
    </script>
  </body>
</html>
